+++ backend-embedded-graphics/src/widgets/border.rs
@@ -2,12 +2,12 @@ use embedded_graphics::{
     draw_target::DrawTarget,
     pixelcolor::PixelColor,
     prelude::Primitive,
-    primitives::{PrimitiveStyle, PrimitiveStyleBuilder},
+    primitives::{PrimitiveStyle, PrimitiveStyleBuilder, StrokeAlignment},
     Drawable,
 };
 use embedded_gui::{
     widgets::{
-        primitives::background::{Background, BackgroundProperties},
+        border::{Border, BorderProperties},
         Widget,
     },
     WidgetRenderer,
@@ -15,54 +15,64 @@ use embedded_gui::{
 
 use crate::{themes::Theme, EgCanvas, ToRectangle};
 
-pub struct BackgroundStyle<C>
+pub struct BorderStyle<C>
 where
     C: PixelColor,
 {
     color: C,
+    width: u32,
 }
 
-impl<C> BackgroundStyle<C>
+impl<C> BorderStyle<C>
 where
     C: PixelColor,
 {
     fn build_style(&self) -> PrimitiveStyle<C> {
-        PrimitiveStyleBuilder::new().fill_color(self.color).build()
+        PrimitiveStyleBuilder::new()
+            .stroke_alignment(StrokeAlignment::Inside)
+            .stroke_color(self.color)
+            .stroke_width(self.width)
+            .build()
     }
 }
 
-impl<C> Default for BackgroundStyle<C>
+impl<C> Default for BorderStyle<C>
 where
     C: Theme,
 {
     fn default() -> Self {
         Self {
-            color: C::BACKGROUND_COLOR,
+            color: C::BORDER_COLOR,
+            width: 1,
         }
     }
 }
 
-impl<C> BackgroundProperties for BackgroundStyle<C>
+impl<C> BorderProperties for BorderStyle<C>
 where
     C: PixelColor,
 {
     type Color = C;
 
-    fn set_background_color(&mut self, color: Self::Color) {
+    fn get_border_width(&self) -> u32 {
+        self.width
+    }
+
+    fn set_border_color(&mut self, color: Self::Color) {
         self.color = color;
     }
 }
 
-// TODO: draw target should be clipped to widget's bounds, so this can be restored to Background
-impl<W, C, DT> WidgetRenderer<EgCanvas<DT>> for Background<W, BackgroundStyle<C>>
+// TODO: draw target should be clipped to widget's bounds, so this can be restored to Border
+impl<W, C, DT> WidgetRenderer<EgCanvas<DT>> for Border<W, BorderStyle<C>>
 where
     W: Widget + WidgetRenderer<EgCanvas<DT>>,
     C: PixelColor,
     DT: DrawTarget<Color = C>,
-    BackgroundStyle<C>: BackgroundProperties,
+    BorderStyle<C>: BorderProperties,
 {
     fn draw(&self, canvas: &mut EgCanvas<DT>) -> Result<(), DT::Error> {
-        let style = self.background_properties.build_style();
+        let style = self.border_properties.build_style();
 
         self.bounding_box()
             .to_rectangle()
