+++ backend-embedded-graphics/src/widgets/mod.rs
@@ -1,6 +1,7 @@
+pub mod background;
+pub mod border;
 pub mod graphical;
 pub mod label;
-pub mod primitives;
 pub mod scroll;
 pub mod text_block;
 pub mod text_box;
