--- backend-embedded-graphics/src/widgets/primitives/mod.rs
@@ -1,2 +1,0 @@
-pub mod background;
-pub mod border;
