+++ examples/calculator.rs
@@ -19,12 +19,10 @@ use embedded_gui::{
     input::event::{InputEvent, PointerEvent},
     prelude::*,
     widgets::{
+        fill::{FillParent, Right},
         label::Label,
         layouts::linear::{column::Column, row::Row},
-        primitives::{
-            fill::{FillParent, Right},
-            spacing::Spacing,
-        },
+        spacing::Spacing,
     },
 };
 use heapless::String;
