--- src/widgets/primitives/mod.rs
@@ -1,5 +1,0 @@
-pub mod background;
-pub mod border;
-pub mod fill;
-pub mod spacing;
-pub mod visibility;
