+++ src/widgets/utils/mod.rs
@@ -1,2 +1,4 @@
+//! Widget and data binding utilities
+
 pub mod decorator;
 pub mod wrapper;
